pyproject.toml
[project]
name = "football-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.3.2",
    "espn-api>=0.45.1",
    "python-dotenv>=1.1.1",
]
